+++ lib/spektr/checks/base.rb
@@ -28,6 +28,7 @@ module Spektr
     def warn!(target, check, location, message, confidence = :high)
       full_path = target.is_a?(String) ? target : target.path
       path = full_path.gsub(@app.root, "")
+      return if target.respond_to?(:ignored_at?) && target.ignored_at?(location&.start_line, check)
       return if dupe?(path, location, message)
 
       @app.warnings << Warning.new(path, full_path, check, location, message, confidence)